src/python/requirements.txt
mysql-connector-python
pyserial
playsound